upbit_trading_bot/pyproject.toml
[project]
name = "upbit-trading-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "pyjwt>=2.10.1",
    "requests>=2.32.4",
    "websocket-client>=1.8.0",
]
